+++ CHANGELOG.md
@@ -5,9 +5,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## [Unreleased][Unreleased]
+
+## [13.2.0][13.2.0] - 2016-08-03
 ### Added
-- [Feature] Created Token react component.
-- [Patch] Include an example of a Table React component with a Button within it.
+- [Feature] Created `Token` React component.
+- [Patch] Include an example of a `Table` React component with a `Button` within it.
 
 ### Fixed
 - [Patch] Add expected spacing between the buttons in the React ButtonRow component. (#473)
@@ -456,7 +458,7 @@ This file follows the format suggested by [Keep a CHANGELOG](https://github.com/
 ### Added
 - [Patch] Use semantic versioning in LEGO (#58).
 
-[Unreleased]: https://github.com/optimizely/oui/compare/v13.1.1...HEAD
+[Unreleased]: https://github.com/optimizely/oui/compare/v13.2.0...HEAD
 [0.0.2]: https://github.com/optimizely/oui/compare/v0.0.1...v0.0.2
 [0.0.3]: https://github.com/optimizely/oui/compare/v0.0.2...v0.0.3
 [1.0.0]: https://github.com/optimizely/oui/compare/v0.0.3...v1.0.0
@@ -497,3 +499,4 @@ This file follows the format suggested by [Keep a CHANGELOG](https://github.com/
 [13.0.1]: https://github.com/optimizely/oui/compare/v13.0.0...v13.0.1
 [13.1.0]: https://github.com/optimizely/oui/compare/v13.0.1...v13.1.0
 [13.1.1]: https://github.com/optimizely/oui/compare/v13.1.0...v13.1.1
+[13.2.0]: https://github.com/optimizely/oui/compare/v13.1.1...v13.2.0